KidneyExchange.jl

Documentation for KidneyExchange.jl

KidneyExchange.BP_info — Type

BP_info

Mutable structure where extra information about the branch-and-price execution is stored

Fields

LB::Float64: best primal value

UB::Float64: best dual value

nb_col_root::Int: number of columns generated at root node

KidneyExchange.BP_params — Type

BP_params

Mutable structure where the options of the branch-and-price solver are stored

Fields

optimizer::String: LP and IP solver that will be used to solve the master (default is Cbc for IPs and GLPK for LPs)

verbose::Bool: true if messages are printed during the solution (default = true)

is_pief::Bool: true if the chains are considered in the master model using a position-indexed extended edge formulation (default = false)

fvs::Bool: true if a feedback vertex set is used to reduce the number of graph copies (default = true)

reduce_vertices::Bool: true if we try deleting useless vertices in graph copies (default = true)

is_column_disjoint::Bool: true if we require column disjoint columns at each column generation iteration (default = true)

max_intersecting_columns::Int: true maximum number of generated columns covering each vertex (default = 6)

is_tabu_list::Bool: true if we stop solving a subproblem as soon as it does not produce any positive cost column; this subproblem will be considered again when proving optimality (default = true)

solve_master_IP::Bool: true if we solve the master IP to find feasible solutions (default = true)

time_limit_master_IP::Float64: time limit (seconds) at each solution of the master IP (default = 10.0)

freq_solve_master_IP::Int: number of new columns that must be added in the master IP between two solutions of this IP (default = 1)

restart_for_IP::Bool: true if the root node can be solved twice to generate more columns when the IP master could not prove optimality of the relaxation value (default = true)

KidneyExchange.BP_status — Type

BP_params

Mutable structure where the results of the branch-and-price are stored

Fields

bp_info::BP_info: extra info about the branch-and-price execution

status::String: status of solution: ONGOING, OPTIMAL or TIMELIMIT

objective_value::Float64: objective value of the best solution found

relative_gap::Float64: final relative optimality gap

best_cycles::Vector{Vector{Int}}: selected cycles in the best primal value

best_chains::Vector{Vector{Int}}: selected chains in the best primal value

node_count::Int: total number of branch-and-bound nodes explored during the solution proces

solve_time::Float64: time spent in the solution process; this might be different from the specified time limit even if status is TIME_LIMIT, since parsing and preprocessing is counted in cpu time

KidneyExchange.Graph_copies — Type

Graph_copies

Mutable structure describing the copies of the graph: one copy per vertex or one copy per vertex of a feedback vertex set if the option is set. It is important to note that the copies related to altruist donors always appear first in the list of copies

Fields

sources::Vector{Int}: source of each graph copy

is_vertex_list::Vector{BitVector}: for each copy and each vertex, true if the vertex belongs to the copy

d_to_vstar_list::Vector{Vector{Int}}: in each copy, distance from each vertex to the source

d_from_vstar_list::Vector{Vector{Int}}: in each copy, distance from each vertex to the source

nb_copies::Int: number of graph copies

is_arc_list::Vector{BitVector}: for each copy and each arc index, true if the arc appears in the copy

chain_mip::Model: shared MIP model that will be solved every time a chain subproblem needs to be solved to optimality

KidneyExchange.Instance — Type

Instance

Non-mutable structure describing the instance that is solved

Fields

graph::SimpleDiGraph: graph describing compatibilities between pairs and with altruist donors

vertex_weight::Vector{Float64}: weight of each vertex

edge_weight::Matrix{Float64}: weight of each arc (0 if there is no arc)

pairs::Vector{Int}: indices of the donor/patient pair vertices

altruists::Vector{Int}: indices of the altruist donor vertices

nb_pairs::Int: number of donor/patient pairs

nb_altruists::Int: number of altruist donors

max_cycle_length::Int: maximum length of a feasible exchange cycle

max_chain_length::Int: maximum length of a feasible exchange chain

is_vertex_weighted::Bool: true if all weights are actually on the vertices

KidneyExchange.MIP_params — Type

MIP_params

Mutable structure where the solving options of the compact formulation are stored

Fields

*optimizer::String: LP and IP solver that will be used to solve the master (default is Cbc for IPs and GLPK for LPs)

verbose::Bool: true if messages are printed during the solution (default = true)

model_type::Mip_model: type of MIP compact model that is to be solved (default = HPIEF)

fvs::Bool: true if a feedback vertex set is used to reduce the number of graph copies (default = true)

reduce_vertices::Bool: true if we try deleting useless arcs in graph copies (default = true)

reduce_arcs::Bool: true if we try deleting useless arcs in graph copies (default = true)

symmetry_break::Bool: true if the MIP model is modified to reduce the number of optimal solutions (default = true)

KidneyExchange.Mip_model — Type

Mip_model

Enumerated structure specifying the considered formulation when solving a compact formulation of the kidney exchange problem

Enumerate values

HPIEF: Hybrid position-indexed extended formulation

EXTENDED_EDGE: Cycles are handled with an extended edge formulation and chains are handled with position-indexed variables

CYCLE_CUT: Cycles are handled with an extended edge formulation and chains are handled with cycle cuts to avoid long cycles

RELAXED_ARC: Arc formulation where the size of cycles and chains are not considered

KidneyExchange.PoolGenerator — Type

Compatibility graph generator based on the following paper: Increasing the Opportunity of Live Kidney Donation by Matching for Two and Three Way Exchanges. S. L. Saidman, Alvin Roth, Tayfun Sonmez, Utku Unver, Frank Delmonico. Transplantation, Volume 81, Number 5, March 15, 2006.

This is known colloquially as the "Saidman Generator".

KidneyExchange.Bellman_Ford_cycle_search — Method

Bellman_Ford_cycle_search

Bellman-Ford style search for one positive cost cycle

Arguments

graph::SimpleDiGraph : The directed graph with cost on each arc

arc_cost::Matrix{Float64}:

source::Int : The local vertex index from which starts the search

K::Int : The maximal length of cycles

pred[k][v] contains u if u is a predecessor of v in a path of length k from source

Return values

cycle::Vector{Int}: the positive cycle found, [] if none

KidneyExchange.MIP_chain_search — Function

MIP_chain_search

IP model with subtour elimination constraints. The constraints are the generalized cutset inequalities (GCS) and they are added dynamically in a row generation algorithm. Refer for instance to the following reference for a presentation of the GCS: Leonardo Taccari (2016).

Arguments

mip::Model: The JuMP model initialized with the flow conservation constraints and the bound on the length of the chain

graph::SimpleDiGraph : The directed graph with cost on each arc

source::Int: Index of the source vertex

is_vertex::BitVector: For each vertex, indicates if it is in the considered subgraph

arc_cost::Vector{Float64}: Matrix of reduced costs of every arc

Return values

is_positivie_chain::Bool: True if a positive chain was found

chain::Vector{Int}: Positive chain that was found

KidneyExchange.SparseUNOSSaidmanPoolGenerator — Method

A tweak to the published Saidman generator; distributions VERY ROUGHLY mimic the UNOS pool as of April 15, 2013. Data taken from the KPD Work Group Data Analysis - CMR - June 2013 report. @author John P. Dickerson

KidneyExchange.activate_branching_constraints — Method

activate_branching_constraints

Activate the all the branching constraints corresponding to a given node of the branch-and-price enumeration tree.

KidneyExchange.bfs — Function

bfs

Breadth-first search to calculate the shortest path in terms of number of arcs from source to other vertices of graph g. The search stops when it is at level K even if there is vertex to be visited. Because we are trying to find cycles of at most K length, vertices of level >=K are never be considered in cycles.

Input parameters

g::AbstractGraph{T} : The graph

source::Int : The source vertex

K::Int: The maximum number of arcs

vertex_in_subgraph::Array{Bool}: Table indicating for each vertex if it is in the subgraph under consideration

Output parameters

dists::Vector{Float64}: shortest distance from source to vertices

KidneyExchange.bfs_reverse — Function

bfs_reverse

Breadth-first search in the graph where arcs are reversed

Input parameters

g::AbstractGraph{T} : The graph

source::Int : The source vertex

K::Int: The maximum number of arcs

vertex_in_subgraph::Array{Bool}: Table indicating for each vertex if it is in the subgraph under consideration

Output parameters

dists::Vector{Float64}: shortest distance from source to vertices

KidneyExchange.branch_and_price — Method

branch_and_price

Core function of the KEP solution with branch-and-price. It requires a parsed instance and the description of the graph copies. The column generation model is that of Riazcos-Alvarez et al (2020), but the many improvements have been added, in particular in the solution of the subproblem.

Arguments

instance::Instance: The parsed instance that is to be solved, it contains the KEP graph and the bounds on the length of covering cycles and chains

subgraphs::Graph_copies: Description of the graph copies of the extended edge formulation

bp_params::BP_params: solution parameters of the branch-and-price

timer::TimerOutput: a timer that will provide detail of where the computational time was spent during the branch-and-price

time_limit::Float64: time limit of the algorithm, including parsing and prepreprocessing times

#Output parametes

bp_status::BP_status: Structure containing every relevant information on the execution of the algorithm (including the optimal solution)

KidneyExchange.branch_on_arc — Function

branch_on_arc

Update the branch-and-bound tree with two new nodes by branching on the given arc with the specified branching (only on master problem or both in master and in subproblem)

Input parameters

arc_to_branch::Pair{Int,Int} : The arc to be branched

master::Model: Master model where the branching constraints are added

is_cg_branching::Bool: True if the branching impacts the subproblem of the coluln generation, false if it impacts only the master problem

tree::Vector{TreeNode}: Branch-and-bound tree

current_node::TreeNode: Branch-and-bound node currently treated

column_pool::Vector{Column}: Pool of all columns in current master problem

KidneyExchange.branch_on_vertex — Function

branch_on_vertex

Update the branch-and-bound tree with two new nodes by branching on the given vertex.

Input parameters

vertex_to_branch::Int : The arc to be branched

master::Model: Master model where the branching constraints are added

tree::Vector{TreeNode}: Branch-and-bound tree

current_node::TreeNode: Branch-and-bound node currently treated

column_pool::Vector{Column}: Pool of all columns in current master problem

KidneyExchange.build_hpief_mip — Function

build_hpief_mip(instance, subgraphs, params; maxtime = 600)

A compact MIP formulation originally proposed by John P. Dickerson, David F. Manlove, Benjamin Plaut, Tuomas Sandholm, James Trimble (2016). The hybrid position indexed edge formulation (HPIEF) handles cycles and chains in the same formulation.

KidneyExchange.build_reduced_extended_edge_mip — Function

build_reduced_extended_edge_mip(instance, subgraphs, params; maxtime = 600)

A compact MIP formulation originally proposed by Miguel Constantino, Xenia Klimentova, Ana Viana, Abdur Rais (2013) for the cycles-only variant of the problem. Here it is adapted to the graph copies based on FVS and chains are considered with position-indexed variables.

KidneyExchange.check_used_vertices — Method

check_used_vertices

Gives a set of indices of cycle in cycles who have common vertices with cycle

Arguments

cycle::Array{Int,1}: The vertices of cycle

cycles::Array{Array{Int,1},1}: The set of cycles

#Output parametes

redundant_cycle_indices::Array{Int,1}: the set of indices of cycle in cycles who have common vertices with cycle

KidneyExchange.compute_arc_flow — Method

compute_arc_flow

Calculates the relaxed value of the decision variable x(i,j) of arc (i->j) in A. x(i,j) whose value >0 is stored in x::Dict{Pair{Int,Int}, Float64}

Arguments

mastersol::Vector{Float64}: solution value of the master problem, ie: y[c] for c in cycles

node_columns::Vector{Column}: The corresponding cycles of current node

Return values

x:: Dict{Pair{Int,Int}, Float64}: value of x_(i,j) of arc (i->j)

KidneyExchange.create_chain_mip — Method

create_chain_mip

Initialize the MIP model with cycle constraint generation for the optimal search of positive cost chains. Only one model is created for every copy to save a great amount of initialization time and memory. The model will then need to be modified for each graph copy to keep only the vertices of the graph and select the right source vertex

Arguments

graph::SimpleDiGraph : The directed graph with cost on each arc

L::Int: The maximal length of chains

optimizer::String: Name of the MIP sover

time_limit::Real: Time limit of the solver

gurobi_env: Gurobi environment if Guroib is used (avoids many messages from the solver)

ln(" . solution found: ", arcs, ", objective value: ", objectivevalue(mip)) println(" . cost of other vertices = ", [vertexcost[e[2]] for e in arcs]) println(" . the mip search did not find any positive cost chain") end

Return values

mip::Model: Initial JuMP model for the search of a positive chain

KidneyExchange.deactivate_branching_constraints — Method

deactivate_branching_constraints

Deactivate the all the branching constraints corresponding to a given node of the branch-and-price enumeration tree.

KidneyExchange.drawPatientBlood_type — Method

Draws a random patient's blood type from the US distribution

@return Blood_type.{O,A,B,AB}

KidneyExchange.generateAltruist — Method

Random rolls an altruistic donor (donor with no attached patient) @param ID unique identifier for the vertex @return altruistic donor vertex KPDVertexAltruist

KidneyExchange.generatePair — Method

Randomly rolls a patient-donor pair (possibly compatible or incompatible) @param ID unique identifier for the vertex @return a patient-donor pair KPDVertexPair

KidneyExchange.generatePraIncompatibility — Method

Randomly generates CPRA (Calculated Panel Reactive Antibody) for a patient-donor pair, using the Saidman method. If the patient is the donor's wife, then CPRA is increased. @param isWifePatient is the patent the wife of the donor? @return scaled CPRA double value between 0 and 1.0

KidneyExchange.generate_abraham_benchmark — Method

generate_abraham_benchmark

Generate a benchmark using the same method as that used in David J. Abraham, Avrim Blum, Tuomas Sandholm (2007)

Abraham, David J., Avrim Blum, et Tuomas Sandholm. "Clearing Algorithms for Barter Exchange Markets: Enabling Nationwide Kidney Exchanges". In Proceedings of the 8th ACM Conference on Electronic Commerce, 295‑304. EC ’07. New York, NY, USA: ACM, 2007. https://doi.org/10.1145/1250910.1250954.

KidneyExchange.generate_complete_benchmark — Method

generate_complete_benchmark

Generate all the instances that are used in addition to the PrefLib to assess the solution methods of this package in the article describing the package.

add citation later*

The instances are stored in the subdirectories heterogeneous, sparse and saidman of the data directory. Beware that if using the package with Pkg.add(), the data directory is in the directory where the package is stored.

KidneyExchange.generate_heterogeneous_instance — Method

generate_heterogeneous_instance

Generate a KEP dataset and write in two text files with extensions .dat and .wmd; the graph generator is based on the following article: Kidney Exchange in Dynamic Sparse Heterogeneous Pools. Itai Ashlagi, Patrick Jaillet, Vahideh H. Manshadi. EC-2013. (Extended abstract.)

Arguments

nb_pairs::Int: number of pairs of incompatible donor and patient

nb_altruists::Int: number of altruist donors

index::Int: index of the dataset, which will appear in the filename

KidneyExchange.generate_heterogeneous_kep_graph — Function

Compatibility graph generator based on the following paper: Kidney Exchange in Dynamic Sparse Heterogeneous Pools. Itai Ashlagi, Patrick Jaillet, Vahideh H. Manshadi. EC-2013. (Extended abstract.) @author John P. Dickerson

KidneyExchange.generate_saidman_instance — Method

generate_saidman_instance

Generate a KEP dataset and write in two text files with extensions .dat and .wmd. This generator is usually refererred to as the "Saidman Generator".

Arguments

nb_pairs::Int: number of pairs of incompatible donor and patient

nb_altruists::Int: number of altruist donors

index::Int: index of the dataset, which will appear in the filename

Reference

Susan L. Saidman, Alvin E. Roth, Tayfun Sonmez, M. Utku Unver, Francis L. Delmonico (2014)

"Increasing the Opportunity of Live Kidney Donation by Matching for Two and Three Way Exchanges". S. L. Saidman, Alvin Roth, Tayfun Sonmez, Utku Unver, Frank Delmonico. Transplantation, Volume 81, Number 5, March 15, 2006.

KidneyExchange.generate_sparse_unos_instance — Method

generate_sparse_unos_instance

Generate a KEP dataset and write in two text files with extensions .dat and .wmd. This generator is obtained by applying the Saidman generator with small modifications in the probabilities of compatibility to roughtly mimic the UNOS pool as of April 15, 2013. The corresponding data was taken from the KPD Work Group Data Analysis - CMR - June 2013 report. Seel also the following article for the motivation and description of this generator: "Optimizing Kidney Exchange with Transplant Chains: Theory and Reality". John P. Dickerson, Ariel D. Procaccia, Tuomas Sandholm; Proceedings of AAMAS; 2012

Arguments

nb_pairs::Int: number of pairs of incompatible donor and patient

nb_altruists::Int: number of altruist donors

index::Int: index of the dataset, which will appear in the filename

KidneyExchange.get_branching_arc — Method

get_branching_arc

Find a fractional arc to branch. The fractional arc closest to 0.5 will be selected to branch, if there is no fractional arc in the solution

Input parameters

column_flow::Dict{Pair{Int,Int}, Float64} : The dictionary containing the nonzero flow on each arc due to the selection of columns

pief_flow::Dict{Pair{Int,Int}, Float64} : The dictionary containing the flow on each arc from the pief model for chain search (if applicable)

Output parameters

arc_to_branch::Pair{Int,Int} : The arc to be branched

is_cg_branching::Bool: True if the branching impacts the subproblem of the coluln generation, false if it impacts only the master problem

KidneyExchange.get_branching_vertex — Method

get_branching_vertex

Find a fractional vertex cover to branch. The fractional vertex closest to 0.5 will be selected to branch

Input parameters

column_flow::Dict{Pair{Int,Int}, Float64} : The dictionary containing the nonzero flow on each arc due to the selection of columns

pief_flow::Dict{Pair{Int,Int}, Float64} : The dictionary containing the flow on each arc from the pief model for chain search (if applicable)

Output parameters

vertex_to_branch::Int : The vertex to be branched on, nothing if none was found

KidneyExchange.get_feasible_solution — Method

get_feasible_solution

Extract a integer feasible solution from the fractional solution by conserving a set of vertex-disjoint cycles or chains

Arguments

fractional_solution::Array{Array{Float64,1},1}: The set of value of variables indicating whether or not cycles and chains are selected

node_columns::Array{Array{Column,1},1}: The set of cycles and chains associated with variables

#Output parametes

feas_val::Real: The objective value of the feasible solution

columns::Array{Array{Int,1},1}: The set of selected cycles and chains of the feasible solution

KidneyExchange.initialize_column_pool — Function

initialize_column_pool

Initialize the pool of columns for the branch-and-price by enumerating all k-cycles up to an input maximum value of k.

Input parameters

*graph::SimpleDiGraph: The KEP graph *column_pool::Vector{Column}: Pool of columns where initial columns will be pushed *max_cycle_length::Int: Maximum length of the cycles to be enumerated at initialization (0 if initialization is deactivated)

Output parameters

column_pool::Vector{Column} : set of columns initially added to the master problem

KidneyExchange.initialize_master_IP — Function

initialize_master_IP

Initialization of the master problem with integer columns

Arguments

instance::Instance: The parsed instance that is to be solved, it contains the KEP graph and the bounds on the length of covering cycles and chains.

column_pool::Vector{Column}: the set of initial columns of the master

bp_params::BP_params: parameters of the branch-and-price

time_limit::Float64: time limit for each solution of the master relaxation

Return values

master::Model: the model of the restricted master problem

KidneyExchange.isDonorSpouse — Method

Draws a random spousal relationship between donor and patient @return true if willing donor is patient's spouse, false otherwise

KidneyExchange.isPatientFemale — Method

Draws a random gender from the US waitlist distribution @return true if patient is female, false otherwise

KidneyExchange.isPositiveCrossmatch — Method

Random roll to see if a patient and donor are crossmatch compatible @param pr_PraIncompatibility probability of a PRA-based incompatibility @return true is simulated positive crossmatch, false otherwise

KidneyExchange.node_master — Function

node_master

Initialization of the restricted master problem

Arguments

instance::Instance: The parsed instance that is to be solved, it contains the KEP graph and the bounds on the length of covering cycles and chains.

column_pool::Vector{Column}: the set of initial columns of the master

bp_params::BP_params: parameters of the branch-and-price

time_limit::Float64: time limit for each solution of the master relaxation

Return values

master::Model: the model of the restricted master problem

KidneyExchange.preprocess_graph_copies — Function

preprocess_graph_copies

Get the data of each graph copy used in the subproblems of the column generation. The graph copies are charaterized by a source vertex and the set of vertices (optionally set of arcs) kept in the copy. Each altruist donor is the source of one graph copy. Preprocessing the vertices that can be reached with a given maximum number of arcs does not bring any improvement to the subproblem solution, so we do nothing at this stage.

For the other copies, either each donor/patient pair is the source of one copy, or we compute a feedback vertex set (FVS) that heuristically minimizes the cardinality of the set. Each vertex of the FVS will then be the source of one copy in the column generation subproblems. In both cases, we then compute distances from and to the source in each subgraph. These will be used in the Bellman-Ford search at each solution of the subproblems.

Input parameters

instance::Instance : The structure describing the characteristics of an instance including the compatibility graph

reduce_arcs::Bool : A boolean parameter indicating whether or not arcs that are not useful in a copy should be eliminated from it (default value is false)

reduce_vertices::Bool : A boolean parameter indicating whether or not vertices that are not useful in copy should be eliminated from it (default value is true)

fvs::Bool : True if the copies correspond to an FVS, false if there is one copy per donor/patient pair

Output parameters

subgraphs::Graph_copies : The structure describing the preprocessed data of the graph copies

KidneyExchange.print_and_check_solution — Function

print_and_check_solution

Function that prints the main characteristics of a solution and check that it satisfies all teh constraints

Input parameters

cycles::Vector{Vector{Int}}: list of cycles in the solution to display

chains::Vector{Vector{Int}}: list of chains in the solution to display

instance::Instance: KEP instance whose solution it is

verbose::Bool: true if the characteristics of the solution are printed, false if only the verification need be done

Output parameters: None

KidneyExchange.process_node — Method

process_node

Columns generation of the node

#Input parameter

tree_node::TreeNode: Branch-and-bound node to solve with column generation

#Output parameter

column_flow::Dict{Pair{Int,Int}, Float64}: value of x_(i,j) of arc (i->j)`

KidneyExchange.read_kep_file — Method

read_kep_file

Contruct a KEP graph from a .wmd and a .dat input files

Parameters

wmd_file::String : Absolute path of the .wmd file.

dat_file::String : Absolute path of the .dat file.

KidneyExchange.relaxed_arc — Function

relaxed_arc_deterministic

Deterministic relaxed-arc formulation for a KEP problem. No cycle length limitation is imposed.

Parameters

instance::Instance: KEP instance to solve

maxtime::Real=60 : Maximum solving time in seconds

KidneyExchange.solve_with_BP — Function

solve_with_BP

This is the main function to call for an execution of the branch-and-price algorithm on input file with given bounds on the length of covering cycles and chains and given options

Arguments

filename::String: path of the input data files, this should include the name of the files, but not the .dat and .wmd extensions

K::Int: maximum length of cycles

L::Int: maximum length of chains

bp_params::BP_params: solution parameters of the branch-and-price

timer::TimerOutput: a timer that will provide detail of where the computational time was spent during the branch-and-price

time_limit::Float64: time limit of the algorithm, including parsing and prepreprocessing times

#Output parametes

instance::Instance: The parsed instance that is to be solved, it contains the KEP graph and the bounds on the length of covering cycles and chains.

subgraphs::Graph_copies: Description of the graph copies of the extended edge formulation

bp_status::BP_status: Structure containing every relevant information on the execution of the algorithm (including the optimal solution)

KidneyExchange.solve_with_mip — Function

solve_with_mip

This is the main function to call to solve the input instance with given bounds on the length of covering cycles and chains and given options

Arguments

filename::String: path of the input data files, this should include the name of the files, but not the .dat and .wmd extensions

K::Int: maximum length of cycles

L::Int: Maximum length of chains

params::MIP_params: parameters of the MIP model and solver

timer::TimerOutput: a timer that will provide detail of where the computational time was spent during the branch-and-price

time_limit::Float64: time limit of the algorithm, including parsing and prepreprocessing times

#Output parametes

instance::Instance: The parsed instance that is to be solved, it contains the KEP graph and the bounds on the length of covering cycles and chains.

subgraphs::Graph_copies: Description of the graph copies of the extended edge formulation

bp_status::BP_status: Structure containing every relevant information on the execution of the algorithm (including the optimal solution)

KidneyExchange.write_kep_file — Method

write_kep_file

Write a .wmd and a .dat files to store the input KEP graph in the same form as in Preflib. Store the files in ./data directory.

Parameters

kep_graph::SimpleDiGraph: KEP graph to write

edge_weight::Matrix{Float64}: weight of each eadge of the KEP graph

donorBT::Vector{Blood_type}: blood type of the donor of each pair

patientBT::Vector{Blood_type}: blood type of the patient of each pair

wifeP::BitArray: for each pair true iff the donor and patient are married

patientPRA::Vector{Float64}: PRA of the patient of each pair

is_altruist::BitArray: for each pair, true if the donor is an altruist donor

file_name::String: Name of the files (before the extensions)